+++ posthog-android/src/main/java/com/posthog/android/replay/internal/Throttler.kt
@@ -5,37 +5,36 @@ import com.posthog.internal.PostHogDateProvider
 import java.util.concurrent.TimeUnit
 import java.util.concurrent.atomic.AtomicBoolean
 
-internal class Debouncer(
+internal class Throttler(
     private val mainHandler: MainHandler,
     private val dateProvider: PostHogDateProvider,
-    private val debouncerDelayMs: Long,
+    private val throttleDelayMs: Long,
 ) {
     private var lastCall = 0L
-    private val delayNs = TimeUnit.MILLISECONDS.toNanos(debouncerDelayMs)
-    private val hasPendingRunnable = AtomicBoolean(false)
+    private val delayNs = TimeUnit.MILLISECONDS.toNanos(throttleDelayMs)
+    private val isThrottling = AtomicBoolean(false)
 
     /**
      * Debounces the given [runnable] by delaying its execution until [delayNs] has passed since the last call.
      */
     internal fun debounce(runnable: Runnable) {
-        if (lastCall == 0L) {
-            lastCall = dateProvider.nanoTime()
-        }
+        val currentTime = dateProvider.nanoTime()
 
-        val timePassedSinceLastExecution = dateProvider.nanoTime() - lastCall
-        if (timePassedSinceLastExecution >= delayNs) {
-            if (!hasPendingRunnable.get()) {
-                execute(runnable)
-            }
+        // Check if enough time has passed since the last execution
+        val timeSinceLastExecution = currentTime - lastCall
+        if (timeSinceLastExecution >= delayNs) {
+            // Execute immediately if enough time has passed
+            execute(runnable)
         } else {
-            if (!hasPendingRunnable.getAndSet(true)) {
+            // If already throttling, ignore additional calls
+            if (!isThrottling.getAndSet(true)) {
                 mainHandler.handler.postDelayed({
                     try {
                         execute(runnable)
                     } finally {
-                        hasPendingRunnable.set(false)
+                        isThrottling.set(false) // Reset throttling after delay
                     }
-                }, debouncerDelayMs)
+                }, throttleDelayMs)
             }
         }
     }