+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- chore: change screenshot image type from JPEG to WEBP ([#211](https://github.com/PostHog/posthog-android/pull/211))
+
 ## 3.9.3 - 2024-11-26
 
 - no user facing changes